requirements.txt
langchain==1.2.10
langchain-core==1.2.16
langchain-community==0.4.1
langchain-google-genai==4.2.1

google-generativeai==0.8.5
google-genai==1.65.0

tavily-python==0.7.22

python-dotenv==1.0.0
requests>=2.32.0